pyproject.toml
[tool.poetry]
name = "photo-bank"
version = "0.1.0"
description = ""
authors = ["a.groza <a.groza.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = {extras = ["standart"], version = "^0.32.0"}
pydantic = {extras = ["email"], version = "^2.9.2"}
pydantic-settings = "^2.6.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
aiosqlite = "^0.21.0"
python-telegram-bot = "^22.0"
jinja2 = "^3.1.6"
Pillow = "^11.2.1"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

